+++ apps/common-app/runtime-tests/reanimated/suites.ts
@@ -5,6 +5,12 @@ type Describe = (name: string, buildSuite: () => void) => void;
 
 export const REANIMATED_TEST_SUITES: RuntimeTestSuite[] = [
   {
+    testSuiteName: 'skew',
+    importTest: () => {
+      require('./tests/props/skew.test');
+    },
+  },
+  {
     testSuiteName: 'animations',
     importTest: () => {
       const { describe } = require('../ReJest/RuntimeTestsApi') as {