+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/common/transforms/TransformOp.h
@@ -18,6 +18,7 @@ enum class TransformOp : uint8_t {
   SkewX,
   SkewY,
   Matrix,
+  Skew,
 };
 
 TransformOp getTransformOperationType(const std::string &property);