+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/interpolation/transforms/TransformOperationInterpolator.cpp
@@ -208,6 +208,7 @@ template class TransformOperationInterpolator<ScaleYOperation>;
 // Skew operations
 template class TransformOperationInterpolator<SkewXOperation>;
 template class TransformOperationInterpolator<SkewYOperation>;
+template class TransformOperationInterpolator<SkewOperation>;
 
 // Translate operations (resolvable)
 template class TransformOperationInterpolator<TranslateXOperation>;